Add files via upload
I have made this website using ReactJS, where the hooks like useState and useEffect is been used. The fetching of the products were done by the axios method, the axios package were being called to set the data inside of the empty array hook and then it is been returned as list by doing mapping on it, where mapping state is being filtered by the .filter method and hence the search condition, uses the same.